[User Tests] Code structure update     - Infra component file is split on separate components     - test_quantizer to more specific test_tensor_quantizer
