+++ user_tests/test_components/test_quality_metrics.cc
@@ -7,7 +7,7 @@
 *
 */
 
-#include "test_infra.h"
+#include "test_quality_metrics.h"
 
 // Standard asserts should be intentionally turned-on by defenition of TEST_DEBUG.
 #if !defined(TEST_DEBUG)
@@ -19,7 +19,6 @@
 #include <algorithm>
 #include <memory>
 
-#include "tensor_transform.h"
 #include "tests_aux.h"
 #include "mli_api.h"
 
@@ -188,196 +187,6 @@ bool quality_metrics::calculate_metrics(const mli_tensor& pred_tsr, const tensor
     quant_error_percentage_ = metrics_quant.noise_vec_length / std::max(metrics_pred.noise_vec_length, eps) * 100.f;
     
     return true;
-}
-
-
-
-//===============================================================================================
-//
-// Methods and data of the Module to calculate and handle CRC32 sum.
-//
-//===============================================================================================
-
-// Half-byte lookup table
-//=========================================
-const uint32_t crc32_calc::crc32_lookup_table_[16] = {
-  0x00000000,0x1DB71064,0x3B6E20C8,0x26D930AC,0x76DC4190,0x6B6B51F4,0x4DB26158,0x5005713C,
-  0xEDB88320,0xF00F9344,0xD6D6A3E8,0xCB61B38C,0x9B64C2B0,0x86D3D2D4,0xA00AE278,0xBDBDF21C
-};
-
-
-// Constructors
-//================================
-crc32_calc::crc32_calc()
-    : crc32_sum_(0x00000000)
-    , valid_crc32_(false)
-{}
-crc32_calc::crc32_calc(uint32_t init_val)
-    : crc32_sum_(init_val)
-    , valid_crc32_(true)
-{}
-
-// Reset instance state
-//================================
-void crc32_calc::reset() {
-    crc32_sum_ = 0x00000000;
-    valid_crc32_ = false;
-}
-void crc32_calc::reset(uint32_t new_val) {
-    crc32_sum_ = new_val;
-    valid_crc32_ = true;
-}
-
-// Get accumulated CRC sum
-//================================
-uint32_t crc32_calc::get() const {
-    return crc32_sum_;
-}
-
-// Get status of CRC instance
-//================================
-bool crc32_calc::is_valid() const {
-    return valid_crc32_;
-}
-
-// Accumulate CRC using tensor data
-//=========================================
-// Note: It was planned to calculate CRC only using valid values of tensor via proper 
-// handling of memstrides. But as we have an opportunity to initialize memory before usage
-// it found out that calculating CRC over all tensor's data container (pointer + capacity)
-// is beneficial as we can track memory corruption of inner tensor data ("holes" that excluded by memstride)
-// TODO: add an optional flag to calculated CRC on valid data, if needed
-uint32_t crc32_calc::operator()(const mli_tensor& in) {
-    const int8_t* current = in.data.mem.pi8;
-    uint32_t length = in.data.capacity;
-
-    return (*this)(current, length);
-}
-
-// Accumulate CRC using array
-//=========================================
-uint32_t crc32_calc::operator()(const int8_t* in, uint32_t size) {
-    if (in != nullptr && size != 0) {
-        uint32_t crc = ~crc32_sum_;  // same as previousCrc32 ^ 0xFFFFFFFF
-
-        while (size-- != 0) {
-            const uint8_t current_val = static_cast<uint8_t>(*in);
-            crc = crc32_lookup_table_[(crc ^ current_val) & 0x0F] ^ (crc >> 4);
-            crc = crc32_lookup_table_[(crc ^ (current_val >> 4)) & 0x0F] ^ (crc >> 4);
-            in++;
-        }
-
-        crc32_sum_ = ~crc; // same as crc ^ 0xFFFFFFFF
-        valid_crc32_ = true;
-    }
-    return crc32_sum_;
-}
-
-
-//=======================================================================
-//
-// Module to handle and check externally allocated memory for test needs
-//
-//=======================================================================
-// Parametrized constructor
-//=========================================
-memory_manager::memory_manager(int8_t* memory, uint32_t mem_size)
-    : source_memory_(memory)
-    , source_mem_size_(mem_size)
-    , allocated_memory_start_(nullptr)
-    , allocated_mem_size_(0)
-    , head_mem_crc_()
-    , tail_mem_crc_()
-{}
-
-// Allocate memory of exact size
-//=========================================
-mli_data_container memory_manager::allocate_memory(uint32_t size, uint32_t fill_pattern) {
-    mli_data_container ret_val{ 0 };
-    if (source_memory_ == nullptr || allocated_memory_start_ != nullptr || source_mem_size_ < size )
-        return ret_val;
-
-    // Fill the whole memory region with a pre-defined pattern
-    int pattern_byte = 0;
-    for (uint32_t idx = 0; idx < source_mem_size_; ++idx, ++pattern_byte) {
-        pattern_byte = pattern_byte % sizeof(fill_pattern);
-        int shift = (sizeof(fill_pattern) - 1 - pattern_byte) * 8;
-        source_memory_[idx] = static_cast<int8_t>((fill_pattern >> shift) & 0xFF);
-    }
-
-    if (size == source_mem_size_) {
-        // If the whole memory region is requested, there is no need to keep valid CRC32 for head and tail
-        // Return the whole memory
-        head_mem_crc_.reset();
-        tail_mem_crc_.reset();
-        allocated_memory_start_ = source_memory_;
-        allocated_mem_size_ = source_mem_size_;
-    } else {
-        // otherwise, we need to return middle sub-region and keep CRC32 checksums for head and tail 
-        const uint32_t head_size = (source_mem_size_ - size) / 2;
-        const uint32_t tail_size = source_mem_size_ - head_size - size;
-        allocated_memory_start_ = source_memory_ + head_size;
-        allocated_mem_size_ = size;
-        head_mem_crc_(source_memory_, head_size);
-        tail_mem_crc_(allocated_memory_start_ + allocated_mem_size_, tail_size);
-    }
-
-    ret_val.capacity = allocated_mem_size_;
-    ret_val.mem.pi8 = allocated_memory_start_;
-    return ret_val;
-}
-
-
-// Allocate memory according to quantizer requirements
-//===================================================
-mli_data_container memory_manager::allocate_memory(const tensor_quantizer& quant_unit, uint32_t fill_pattern) {
-    // First get memory requirements from quantizer
-    mli_data_container empty_container{ 0 };
-    mli_tensor tensor_with_requirements = quant_unit.get_not_quantized_tensor(empty_container);
-
-    // Sort out cases with bad or complete tensors
-    if (tensor_quantizer::validate_tensor(tensor_with_requirements) != tensor_quantizer::kIncompleteMem)
-        return empty_container;
-
-    // calculate how much memory is needed
-    uint32_t required_size = tensor_with_requirements.data.capacity;
-    if (tensor_with_requirements.el_type == MLI_EL_SA_8 || tensor_with_requirements.el_type == MLI_EL_SA_32) {
-        required_size += tensor_with_requirements.el_params.sa.scale.capacity;
-        required_size += tensor_with_requirements.el_params.sa.scale_frac_bits.capacity;
-        required_size += tensor_with_requirements.el_params.sa.zero_point.capacity;
-    }
-
-    return allocate_memory(required_size);
-}
-
-// Mark memory as unused 
-//===================================================
-void memory_manager::return_memory(){
-    head_mem_crc_.reset();
-    tail_mem_crc_.reset();
-    allocated_memory_start_ = nullptr;
-    allocated_mem_size_ = 0;
-}
-
-// Check that Head and tail regions are not corrupted
-//===================================================
-bool memory_manager::is_memory_corrupted() const {
-    if (source_memory_ == nullptr || allocated_memory_start_ == nullptr ||
-            allocated_memory_start_ <= source_memory_  ||
-            allocated_mem_size_ >= source_mem_size_)
-        return false;
-    
-    const uint32_t head_size = static_cast<uint32_t>(allocated_memory_start_ - source_memory_);
-    const uint32_t tail_size = source_mem_size_ - (head_size + allocated_mem_size_);
-    crc32_calc head_current_crc, tail_current_crc;
-    head_current_crc(source_memory_, head_size);
-    tail_current_crc(allocated_memory_start_ + allocated_mem_size_, tail_size);
-
-    assert(head_mem_crc_.is_valid() == head_current_crc.is_valid()); 
-    assert(tail_mem_crc_.is_valid() == tail_current_crc.is_valid());
-    bool is_corrupted = head_mem_crc_.get() != head_current_crc.get();
-    is_corrupted |= tail_mem_crc_.get() != tail_current_crc.get();
-    return is_corrupted;
 }
 
 } // namespace tst
